requirements.txt
alembic==1.8.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Authlib==0.15.6
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.11.0
backports.zoneinfo==0.2.1
bcrypt==3.2.2
bidict==0.22.0
billiard==3.6.4.0
blinker==1.5
boto3==1.20.54
botocore==1.23.54
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
Django==4.1.4
django-celery-beat==2.4.0
django-fabric==2.1.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-yasg==1.21.4
email-validator==1.3.0
eventlet==0.33.0
fabric==2.7.1
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
httpagentparser==1.9.5
idna==2.10
importlib-metadata==5.1.0
importlib-resources==5.10.1
inflection==0.5.1
invoke==1.7.3
isodate==0.6.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.3
kombu==5.2.4
ldap3==2.9.1
Mako==1.2.4
MarkupSafe==2.1.1
msal==1.20.0
msal-extensions==0.3.1
msrest==0.7.1
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==22.0
paramiko==2.12.0
passlib==1.7.4
pathlib2==2.3.7.post1
pgadmin4==6.17
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
portalocker==2.6.0
prompt-toolkit==3.0.36
psutil==5.9.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyotp==2.7.0
pyrsistent==0.19.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
redis==4.4.0
requests==2.25.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.5.2
simplejson==3.18.0
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.45
sqlparse==0.4.3
sshtunnel==0.4.0
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
tzdata==2022.7
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.11.0
